requirements.txt
jupyter
pandas
tqdm
transformers
tensorflow
sentence-transformers
spacy
umap-learn
matplotlib
BERTopic
torch
cuda-python
